src/webhook-handler/requirements.txt
fastapi
uvicorn
gunicorn
google-cloud-pubsub==2.21.5